# Fix Issue 19404 - Optimize std.string.stripRight