Add ASCII fast path for strings and avoid decoding altogether for wstrings. The benchmark was performed with the same inputs as https://github.com/dlang/phobos/pull/6727.

**string benchmark, ldc2:**

| compiler       | function       | data      | time                                |
|----------------|----------------|-----------|-------------------------------------|
| ldc2 -O2       | old_stripRight | ascii     | 1 sec, 396 ms, 375 μs, and 4 hnsecs |
| ldc2 -O2       | new_stripRight | ascii     | 553 ms, 499 μs, and 2 hnsecs        |
| ldc2 -O2       | old_stripRight | non_ascii | 1 sec, 311 ms, 629 μs, and 2 hnsecs |
| ldc2 -O2       | new_stripRight | non_ascii | 1 sec, 394 ms, 103 μs, and 5 hnsecs |

**wstring benchmark, ldc2:**

| compiler       | function       | data        | time                                |
|----------------|----------------|-------------|-------------------------------------|
| ldc2 -O2       | old_stripRight | w_ascii     | 1 sec, 475 ms, 265 μs, and 3 hnsecs |
| ldc2 -O2       | new_stripRight | w_ascii     | 550 ms, 852 μs, and 2 hnsecs        |
| ldc2 -O2       | old_stripRight | w_non_ascii | 1 sec, 76 ms, 60 μs, and 3 hnsecs   |
| ldc2 -O2       | new_stripRight | w_non_ascii | 467 ms, 999 μs, and 6 hnsecs        |

**string benchmark, dmd:**

| compiler       | function       | data      | time                                 |
|----------------|----------------|-----------|--------------------------------------|
| dmd -O -inline | old_stripRight | ascii     | 2 secs, 113 ms, 161 μs, and 2 hnsecs |
| dmd -O -inline | new_stripRight | ascii     | 620 ms, 329 μs, and 8 hnsecs         |
| dmd -O -inline | old_stripRight | non_ascii | 1 sec, 855 ms, 12 μs, and 3 hnsecs   |
| dmd -O -inline | new_stripRight | non_ascii | 1 sec, 951 ms, 677 μs, and 3 hnsecs  |

**wstring benchmark, dmd:**
*The ascii and non-ascii benchmarks use different test data so you shouldn't read anything into the old version taking longer on ascii than non-ascii.*

| compiler       | function       | data        | time                                |
|----------------|----------------|-------------|-------------------------------------|
| dmd -O -inline | old_stripRight | w_ascii     | 2 secs, 86 ms, 815 μs, and 8 hnsecs |
| dmd -O -inline | new_stripRight | w_ascii     | 895 ms, 193 μs, and 7 hnsecs        |
| dmd -O -inline | old_stripRight | w_non_ascii | 1 sec, 535 ms, 802 μs, and 6 hnsecs |
| dmd -O -inline | new_stripRight | w_non_ascii | 677 ms, 969 μs, and 5 hnsecs        |