@n8sh 2018-11-16
It doesn't have any effect on the dstring case: with optimization enabled the same machine code is generated.